update temperature status, notification

+++ LeleGroup/Model/BatasSuhu+CoreDataProperties.swift
@@ -16,10 +16,10 @@ extension BatasSuhu {
         return NSFetchRequest<BatasSuhu>(entityName: "BatasSuhu")
     }
 
-    @NSManaged public var suhuKuningAtas: Int64
-    @NSManaged public var suhuKuningBawah: Int64
-    @NSManaged public var suhuKritisAtas: Int64
-    @NSManaged public var suhuKritisBawah: Int64
+    @NSManaged public var suhuKuningAtas: Double
+    @NSManaged public var suhuKuningBawah: Double
+    @NSManaged public var suhuKritisAtas: Double
+    @NSManaged public var suhuKritisBawah: Double
 
 }
